LayoutTests/platform/win/accessibility/disabled-controls-not-focusable-expected.txt


This test makes sure that disabled form controls are not exposed as focusable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


button
FAIL axElement.isFocusable should be true. Was false.
PASS axElement.isFocusable is false
text
FAIL axElement.isFocusable should be true. Was false.
PASS axElement.isFocusable is false
checkbox
FAIL axElement.isFocusable should be true. Was false.
PASS axElement.isFocusable is false
radio
FAIL axElement.isFocusable should be true. Was false.
PASS axElement.isFocusable is false
submit
FAIL axElement.isFocusable should be true. Was false.
PASS axElement.isFocusable is false
slider
FAIL axElement.isFocusable should be true. Was false.
PASS axElement.isFocusable is false
combobox
FAIL axElement.isFocusable should be true. Was false.
PASS axElement.isFocusable is false
listbox
FAIL axElement.isFocusable should be true. Was false.
PASS axElement.isFocusable is false
textarea
FAIL axElement.isFocusable should be true. Was false.
PASS axElement.isFocusable is false
PASS successfullyParsed is true

TEST COMPLETE
